+++ packages/cli/CHANGELOG.md
@@ -10,6 +10,15 @@
   - More accurately monitors signals (ctrl+c) and shutdowns.
   - Tasks can now be configured with a timeout.
 
+## Unreleased
+
+#### 🐞 Fixes
+
+- Fixed an issue with `moon docker scaffold` where Rust projects in the workspace skeleton would
+  fail to compile as they were missing a lib/main entry point.
+- Fixed an issue with `moon docker prune` where an unknown project type would trigger toolchain
+  errors.
+
 ## 1.22.8
 
 #### 🐞 Fixes